Implement and specialize a `TwoSidedExtend` trait.
This is the the two-sided alternative to `std::iter::Extend`
and supports both `extend_front` and `extend_back`.
Aside from the performance boost,
another major advantage of this change is that
you can extend from an &T iterator if T is Copy.

This is specialized to use optimized implementations where possible,
in order to improve the performance of `extend`.
In particular `std::iter::TrustedLen` is used to eliminate capacity checks,
and we perform a direct `memcpy` when we see `std::slice::Iter`.